Make slice size into a 32 bit quantity -- sizeof does not change

+++ slice.h
@@ -12,7 +12,7 @@
 #include <stdbool.h>
 
 typedef uint8_t  Byte;
-typedef uint16_t Size;
+typedef uint32_t Size;
 
 typedef struct Slice {
     const Byte* ptr;    // pointer to beginning of data
